requirement.txt
discord-py-interactions==4.3.4
discord.py==2.1.1
py-cord==2.3.2
openai==0.25.0
